Fix execution logic and add percentage filed

+++ asura-core/src/main/scala/asura/core/cs/model/AggsItem.scala
@@ -3,7 +3,7 @@ package asura.core.cs.model
 case class AggsItem(
                      val `type`: String,
                      val id: String,
-                     val count: Int,
+                     val count: Long,
                      var sub: Seq[AggsItem] = null,
                      var summary: String = null,
                      var description: String = null,
